requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.25
django-allauth==0.54.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==23.0.0
oauthlib==3.2.2
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.2
requests-oauthlib==2.0.0
sqlparse==0.5.1
